+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.6.0 (April 10, 2020)
+
+FIXES:
+
+* Fixed possible incorrect result for FromFloat64.
+
+FEATURES:
+
+* value: add `Div` and `DivMod` to divide values.
+* value: add `IsZero`.
+
+IMPROVEMENTS:
+
+* added mode tests for `FromFloat64`, `FromString`, JSON unmarshalling.
+* readme: added goreportcard badge.
+
 ## 0.5.0 (March 29, 2020)
 
 FEATURES: